feat: add support for THI events and update related components
- Introduced a new screen for THI events and integrated it into the navigation stack.
- Updated existing event handling to accommodate THI department events.
- Refactored event loading logic to differentiate between student associations and THI departments.
- Added new cards and icons for THI events in the dashboard.
- Updated localization files to include THI events translations.
- Deprecated the old club organizer screen in favor of the new organizer screen for better clarity
