+++ src/app/_layout.tsx
@@ -340,6 +340,13 @@ function RootLayout(): React.JSX.Element {
 					}}
 				/>
 				<Stack.Screen
+					name="(screens)/thi-events"
+					options={{
+						title: t('navigation.thiEvents'),
+						...transparentHeaderStyle
+					}}
+				/>
+				<Stack.Screen
 					name="(screens)/sports"
 					options={{
 						title: t('navigation.sports'),
@@ -495,7 +502,13 @@ function RootLayout(): React.JSX.Element {
 				<Stack.Screen
 					name="(screens)/events/club/[id]"
 					options={{
-						title: t('navigation.club'),
+						headerShown: false
+					}}
+				/>
+				<Stack.Screen
+					name="(screens)/events/organiser/[id]"
+					options={{
+						title: t('navigation.organiser'),
 						...Platform.select({
 							ios: {
 								...presentationMode
