+++ src/components/Cards/index.ts
@@ -3,4 +3,5 @@ export { default as EventsCard } from './events-card'
 export { default as LinkCard } from './link-card'
 export { default as LoginCard } from './login-card'
 export { default as SportsCard } from './sports-card'
+export { default as ThiEventsCard } from './thi-events-card'
 export { default as UpNextCard } from './up-next-card'
