+++ src/components/Error/event-error-view.tsx
@@ -56,13 +56,17 @@ export function EventErrorView({
 	}
 
 	const handleBackToList = (): void => {
-		switch (eventType) {
-			case 'clEvents':
-				router.navigate('/cl-events')
-				break
-			case 'sports':
-				router.navigate('/sports')
-				break
+		if (router.canGoBack()) {
+			router.back()
+		} else {
+			switch (eventType) {
+				case 'clEvents':
+					router.navigate('/cl-events')
+					break
+				case 'sports':
+					router.navigate('/sports')
+					break
+			}
 		}
 	}
