+++ src/components/all-cards.tsx
@@ -6,6 +6,7 @@ import {
 	LinkCard,
 	LoginCard,
 	SportsCard,
+	ThiEventsCard,
 	UpNextCard
 } from './Cards'
 import NewsCard from './Cards/news-card'
@@ -24,6 +25,13 @@ export const AllCards: Card[] = [
 		initial: [USER_STUDENT, USER_EMPLOYEE, USER_GUEST],
 		allowed: [USER_STUDENT, USER_EMPLOYEE, USER_GUEST],
 		card: () => <EventsCard />
+	},
+	{
+		key: 'thiEvents',
+		removable: true,
+		initial: [USER_STUDENT, USER_EMPLOYEE, USER_GUEST],
+		allowed: [USER_STUDENT, USER_EMPLOYEE, USER_GUEST],
+		card: () => <ThiEventsCard />
 	},
 	{
 		key: 'sports',
