+++ src/localization/de/navigation.json
@@ -22,7 +22,7 @@
 		"examDetails": "Prüfungsdetails",
 		"quicklinks": "Quick Links",
 		"dots": "Vorlesungsstatus",
-		"club": "Verein Details",
+		"organiser": "Veranstalter Details",
 		"lecturers": {
 			"title": "Dozenten",
 			"search": "Suche alle Dozenten"
@@ -45,6 +45,7 @@
 		"sports": "Hochschulsport",
 		"roomReport": "Problem mit einem Raum melden",
 		"clEvents": "Campus Life Events",
+		"thiEvents": "THI Events",
 		"neulandMember": "Neuland Mitgliederbereich"
 	},
 	"cards": {
@@ -56,6 +57,7 @@
 			"news": "THI Neuigkeiten",
 			"library": "Bibliothek",
 			"events": "Campus Life Events",
+			"thiEvents": "THI Events",
 			"reimanns": "Reimanns",
 			"canisius": "Canisius Konvikt",
 			"mensa": "Mensa Ingolstadt",
