+++ src/localization/en/navigation.json
@@ -25,7 +25,7 @@
 			"title": "Lecturers",
 			"search": "Search all lecturers"
 		},
-		"club": "Club Details",
+		"organiser": "Organiser Details",
 		"lecturer": "Lecturer Details",
 		"grades": {
 			"title": "Grades",
@@ -45,6 +45,7 @@
 		"sports": "University Sports",
 		"roomReport": "Report Room Issue",
 		"clEvents": "Campus Life Events",
+		"thiEvents": "THI Events",
 		"neulandMember": "Neuland Member Area"
 	},
 	"cards": {
@@ -56,6 +57,7 @@
 			"news": "THI News",
 			"library": "Library",
 			"events": "Campus Life Events",
+			"thiEvents": "THI Events",
 			"reimanns": "Reimanns",
 			"canisius": "Canisius Konvikt",
 			"mensa": "Mensa Ingolstadt",
